Extract helpers.js
