+++ Makefile
@@ -14,7 +14,6 @@ extension:
 	$(MAKE) -C src
 
 EXTENSION_FILES := \
-	src/*.css \
 	src/*.png \
 	src/*.svg \
 	src/fonts/* \
@@ -28,7 +27,8 @@ EXTENSION_FILES := \
 	src/js/background.dist.js \
 	src/js/popup.dist.js \
 	src/js/options.dist.js \
-	src/js/inject.dist.js
+	src/js/inject.dist.js \
+	src/js/helpers.dist.js
 CHROMIUM_FILES := $(patsubst src/%,chromium/%, $(EXTENSION_FILES))
 FIREFOX_FILES  := $(patsubst src/%,firefox/%,  $(EXTENSION_FILES))
