+++ src/popup/popup.js
@@ -2,9 +2,9 @@
 "use strict";
 
 require("chrome-extension-async");
-var TldJS = require("tldjs");
-var sha1 = require("sha1");
-var Interface = require("./interface");
+const sha1 = require("sha1");
+const Interface = require("./interface");
+const helpers = require("../helpers");
 
 run();
 
@@ -27,35 +27,6 @@ function handleError(error, type = "error") {
     errorNode.textContent = error.toString();
     document.body.innerHTML = "";
     document.body.appendChild(errorNode);
-}
-
-/**
- * Get the deepest available domain component of a path
- *
- * @since 3.0.0
- *
- * @param string path        Path to parse
- * @param string currentHost Current hostname for the active tab
- * @return string|null Extracted domain
- */
-function pathToDomain(path, currentHost) {
-    var parts = path.split(/\//).reverse();
-    for (var key in parts) {
-        if (parts[key].indexOf("@") >= 0) {
-            continue;
-        }
-        var t = TldJS.parse(parts[key]);
-        if (
-            t.isValid &&
-            ((t.tldExists && t.domain !== null) ||
-                t.hostname === currentHost ||
-                currentHost.endsWith(`.${t.hostname}`))
-        ) {
-            return t.hostname;
-        }
-    }
-
-    return null;
 }
 
 /**
@@ -98,7 +69,7 @@ async function run() {
                     login: response.files[storeId][key].replace(/\.gpg$/i, ""),
                     allowFill: true
                 };
-                login.domain = pathToDomain(storeId + "/" + login.login, settings.host);
+                login.domain = helpers.pathToDomain(storeId + "/" + login.login, settings.host);
                 login.inCurrentDomain =
                     settings.host == login.domain || settings.host.endsWith("." + login.domain);
                 login.recent =
